moved auth middleware
